chore: Storybook 환경에서 루트 tsconfig 동기화되지 않는 점 해소
